style: InputComment 컴포넌트 생성
