+++ src/app/(afterLogin)/posts/[postId]/layout.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+interface Props {
+  children: ReactNode;
+}
+
+export default function Layout({ children }: Props) {
+  return <div className="min-h-screen flex flex-col justify-between bg-white">{children}</div>;
+}
