+++ src/app/(afterLogin)/posts/[postId]/page.tsx
@@ -1,12 +1,9 @@
-'use client';
-
 import DailyPostView from '@/app/(afterLogin)/posts/_components/DailyPostView';
 import CommentView from '@/app/(afterLogin)/posts/_components/comment/CommentView';
+import InputComment from '@/app/(afterLogin)/posts/_components/input/InputComment';
 import Border from '@/app/_components/common/border/Border';
 import BasicHeader from '@/app/_components/common/header/BasicHeader';
-import InputChat from '@/app/_components/input/InputChat';
 import { DailyPostType } from '@/type';
-import { useState } from 'react';
 
 const DUMMY_DATA: DailyPostType = {
   profileUrl: 'https://avatars.githubusercontent.com/u/49144662?s=400&u=903e697529c3b51f9c69bc3885c8f9be3d754028&v=4',
@@ -94,35 +91,16 @@ const DUMMY_COMMENT = [
   },
 ];
 
-const MAX_LENGTH = 10;
-const PLACEHOLDER = '아들님의 이야기를 적어주세요';
-
 export default function PostPage() {
-  const [inputData, setInputData] = useState<string>('');
-
-  const handleChatChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const input = event.target.value;
-    if (input.length <= MAX_LENGTH) {
-      setInputData(input);
-    }
-  };
-
   return (
-    <div className="min-h-screen flex flex-col justify-between bg-white">
+    <>
       <div>
         <BasicHeader title="게시글" />
         <DailyPostView data={DUMMY_DATA} />
         <Border />
       </div>
       <CommentView data={DUMMY_COMMENT} />
-      <div className="flex items-center justify-center w-full h-[6.0rem] border-bgGrey border-t bg-white">
-        <InputChat
-          value={inputData}
-          maxLength={MAX_LENGTH}
-          placeholder={PLACEHOLDER}
-          onChange={(e) => handleChatChange(e)}
-        />
-      </div>
-    </div>
+      <InputComment />
+    </>
   );
 }
