+++ src/app/globals.css
@@ -24,7 +24,7 @@
     @apply text-black font-700 text-18;
   }
   .header {
-    @apply flex justify-center items-center relative bg-white w-[37.5rem] h-[4.8rem] border-bgGrey border-b;
+    @apply flex sticky top-0 z-10 justify-center items-center bg-white w-[37.5rem] h-[4.8rem] border-bgGrey border-b;
   }
   .dayChip {
     @apply flex items-center justify-center w-[4.8rem] h-[7.6rem] bg-white rounded-[0.6rem] font-body;
